fix: byteArray encoding for less than 31 chars (#1011)
* fix: byteArray encoding for less than 31 chars

* refactor: reuse byteArray utility for typed data hashing

---------
